backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
supabase==2.9.0
python-dotenv==1.0.1

